requirements.txt
# Core Dependencies
streamlit>=1.30.0
streamlit-drawable-canvas>=0.9.3
streamlit-agraph>=0.0.45
streamlit-flow>=0.1.0

# Microsoft Agent Framework
autogen>=0.2.0
semantic-kernel>=0.9.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Azure Services
azure-identity>=1.15.0
azure-ai-inference>=1.0.0
azure-search-documents>=11.4.0
azure-cosmos>=4.5.0
azure-storage-blob>=12.19.0
msal>=1.26.0
msal-extensions>=1.0.0

# LLM Providers
openai>=1.10.0
anthropic>=0.18.0

# Vector Databases
chromadb>=0.4.22
faiss-cpu>=1.7.4
pinecone-client>=3.0.0
qdrant-client>=1.7.0
weaviate-client>=3.25.0

# Data Processing
pandas>=2.1.4
numpy>=1.26.3
python-dotenv>=1.0.0
pyyaml>=6.0.1
jsonschema>=4.20.0

# Authentication
authlib>=1.3.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2
itsdangerous>=2.1.2

# Utilities
requests>=2.31.0
httpx>=0.26.0
pydantic>=2.5.3
tenacity>=8.2.3
tiktoken>=0.5.2

# Testing & Evaluation
pytest>=7.4.3
pytest-asyncio>=0.23.3
pytest-mock>=3.12.0

# Deployment
docker>=7.0.0
python-multipart>=0.0.6
